added permissions for staff and admins
staff can edit posts and profiles; admins can edit and delete posts and profiles

+++ userprofile/views.py
@@ -71,24 +71,32 @@ class UserLogout(View):
 
 
 @login_required(login_url='login')
-def userdelete(request):
-    user = request.user
+def userdelete(request, uname):
+    user = User.objects.get(username=uname)
+    if user != request.user and not request.user.is_superuser:
+        return redirect('home')
     user.delete()
     return redirect('signup')
 
 
 @login_required(login_url='login')
-def useredit(request):
-    user_profile = request.user.userprofile
+def useredit(request, uname):
+    if not User.objects.filter(username=uname).exists():
+        return redirect('home')
+    cur_user = User.objects.get(username=uname)
+    user_profile = UserProfile.objects.get(user=cur_user)
+    if cur_user != request.user and not request.user.is_staff and not request.user.is_superuser:
+        return redirect('home')
     if request.method == 'GET':
         form = UserProfileForm(instance=user_profile, initial=user_profile.__dict__)
     else:
         form = UserProfileForm(request.POST or None, request.FILES or None, instance=user_profile)
         if form.is_valid():
             form.save()
-            return redirect('show profile', str(request.user))
+            return redirect('show profile', str(cur_user))
         else:
             messages.info(request, ":(")
 
-    context = {"profile_form": form}
+    context = {"profile_form": form,
+               "cur_user": cur_user}
     return render(request, template_name="editprofile.html", context=context)
